requirements.txt
pytest==4.3.0
flake8==3.7.7
tox==3.7.0
kubernetes==8.0.1
PyYAML==3.12
jsonschema==3.0.1